backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
numpy==1.26.4
faiss-cpu==1.8.0
GitPython==3.1.43
python-multipart==0.0.9
requests==2.32.3
tree_sitter==0.20.4  # <-- Downgraded for compatibility
tree_sitter_languages==1.10.2
transformers==4.42.4
torch==2.3.1
pytest==8.2.2
pytest-asyncio==0.23.8